Feature/interactive changes (#115)
* create one generateProps method instead of three

* interactive title as component

* fix badge display name

* keep one button to show code, simplify code

* simplify renderInput method, create input components

* move interact to new folder

* move components to corresponding files

* rename styled wrapper

* remove handleShowCode from interact

* add spacing to group form fields

* method renderInput to PropFormField compnent

* add tooltip icon styles

* use PropFormField

* new styles for component name button

* keep text children component visible all the time

* move styled input to propInput component file

* style select and checkbox fields

* fix key on map items

* add text for comment
