+++ packages/styleguide/src/components/Badge/Badge.js
@@ -5,7 +5,7 @@ import { rem } from './../../style/utils';
 
 const propTypes = {
   /** Color as a string from styleguide theme */
-  color: string
+  color: string,
 };
 
 const Badge = ({ children, color = 'black', ...other }) => (
@@ -24,6 +24,7 @@ const StyledBadge = styled.span`
   text-transform: capitalize;
 `;
 
+Badge.displayName = 'Badge';
 Badge.propTypes = propTypes;
 
 export default Badge;
